chore(wmg): move InfoPanel to wmgv2
