+++ frontend/src/views/WheresMyGene/components/GetStarted/style.ts
@@ -8,7 +8,7 @@ import {
 import { LEGEND_MARGIN_BOTTOM_PX } from "../../style";
 import { CELL_INFO_SIDEBAR_WIDTH_PX } from "../../../WheresMyGeneV2/components/CellInfoSideBar/style";
 import { Y_AXIS_CHART_WIDTH_PX } from "../HeatMap/utils";
-import { LEGEND_HEIGHT_PX } from "../InfoPanel/components/Legend/style";
+import { LEGEND_HEIGHT_PX } from "../../../WheresMyGeneV2/components/InfoPanel/components/Legend/style";
 
 export const Header = styled.h1`
   margin-bottom: 12px;
