+++ frontend/src/views/WheresMyGene/components/Main/index.tsx
@@ -32,8 +32,8 @@ import GeneInfoBar from "../../../../components/GeneInfoSideBar";
 import Filters from "../../../WheresMyGeneV2/components/Filters";
 import GetStarted from "../GetStarted";
 import HeatMap from "../HeatMap";
-import InfoPanel from "../InfoPanel";
-import Legend from "../InfoPanel/components/Legend";
+import InfoPanel from "../../../WheresMyGeneV2/components/InfoPanel";
+import Legend from "../../../WheresMyGeneV2/components/InfoPanel/components/Legend";
 import Loader from "../Loader";
 import ScreenTint from "../ScreenTint";
 import { StyledBannerContainer, StyledSidebarDrawer } from "./style";
