+++ frontend/src/views/WheresMyGeneV2/components/Filters/components/ColorScale/index.tsx
@@ -1,7 +1,7 @@
 import { Tooltip } from "@czi-sds/components";
 import questionMarkIcon from "src/common/images/question-mark-icon.svg";
 import { StyledDropdown, Wrapper } from "../common/style";
-import { Label } from "src/views/WheresMyGene/components/InfoPanel/common/style";
+import { Label } from "src/views/WheresMyGeneV2/components/InfoPanel/common/style";
 import { LabelWrapper } from "./style";
 import {
   StyledIconImage,
