+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/common/style.ts
