+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/Methodology/index.tsx
