+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/components/Methodology/style.ts
