+++ frontend/src/views/WheresMyGeneV2/components/InfoPanel/index.tsx
